requirements.txt
aiogram==2.25.1
python-dotenv == 1.0.0
openai==1.1.1
numpy == 1.24.3
soundfile == 0.12.1